+++ example/android/app/src/main/java/com/example/clevertap_plugin_example/MyApplication.java
@@ -9,6 +9,7 @@ public class MyApplication extends FlutterApplication {
     @java.lang.Override
     public void onCreate() {
         CleverTapAPI.setDebugLevel(3);
+        CleverTapAPI.setUIEditorConnectionEnabled(true);
         ActivityLifecycleCallback.register(this);
         super.onCreate();
     }