+++ example/ios/Runner/AppDelegate.m
@@ -2,6 +2,7 @@
 #import "GeneratedPluginRegistrant.h"
 #import "CleverTap.h"
 #import "CleverTapPlugin.h"
+#import "CleverTap+ABTesting.h"
 #import <UserNotifications/UserNotifications.h>
 
 @implementation AppDelegate
@@ -11,7 +12,9 @@
     UNUserNotificationCenter *center = [UNUserNotificationCenter currentNotificationCenter];
     center.delegate = self;
     // Override point for customization after application launch.
+    [CleverTap setUIEditorConnectionEnabled:YES];
     [CleverTap autoIntegrate]; // integrate CleverTap SDK using the autoIntegrate option
+    [CleverTap setDebugLevel:3];
     [[CleverTapPlugin sharedInstance] applicationDidLaunchWithOptions:launchOptions];
     return [super application:application didFinishLaunchingWithOptions:launchOptions];
 }