+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## CHANGE LOG
 
+### Version 1.1.0 (February 27, 2020)
+* Adds support for Dynamic Variables & Native Display
+* Adds support for Google Play Install Referrer Library v1.0
+* Supports CleverTap Android SDK v3.6.4
+* Supports CleverTap iOS SDK v3.7.2
+
 ### Version 1.0.0 (January 20, 2020)
 * Initial Release.
 * Supports CleverTap Android SDK v3.6.1